requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.0.3
smmap==5.0.1
urllib3==2.3.0
